feat: create step2
