+++ src/app/(createADao)/create/component/GovernanceModel/index.css
@@ -1,5 +1,4 @@
 .governance-form {
-
   .TMRWDAO-input-number {
     @apply w-full;
   }
@@ -9,5 +8,12 @@
 
   .form-title-primary {
     @apply text-[20px] leading-[28px] text-neutralTitle font-medium;
+  }
+  .form-list {
+    @apply flex flex-col gap-[50px]
+  }
+  .form-item-label-text {
+    font-family: Montserrat;
+    @apply text-white leading-4
   }
 }
