+++ src/app/(createADao)/create/component/Navigation/index.css
@@ -1,0 +1,8 @@
+.create-dao-navigation {
+  .home-text {
+    font-family: Montserrat;
+  }
+  .create-dao {
+    font-family: Montserrat;
+  }
+}
